feat(backend): add exception handling to UserService

+++ backend/src/main/java/com/example/backend/services/UserService.java
@@ -1,6 +1,7 @@
 package com.example.backend.services;
 
 import com.example.backend.entities.UserEntity;
+import com.example.backend.exceptions.EntityNotFoundException;
 import com.example.backend.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,18 +18,42 @@ public class UserService {
     }
 
     public UserEntity getUserById(int id) {
-        return userRepository.findById(id);
+       UserEntity user = userRepository.findById(id);
+       if (user == null) {
+           throw new EntityNotFoundException("User not found");
+       }
+
+       return user;
     }
 
     public UserEntity saveUser(UserEntity user) {
+        UserEntity existingUser = userRepository.findByName(user.getName());
+        if (existingUser != null) {
+            throw new IllegalStateException("User already exists");
+        }
+
         return userRepository.save(user);
     }
 
     public UserEntity updateUser(UserEntity user) {
+        UserEntity possibleUser = userRepository.findById(user.getId());
+        if (possibleUser == null) {
+            throw new EntityNotFoundException("User not found");
+        }
+        UserEntity existingUser = userRepository.findByNameAndNotId(user.getName(), user.getId());
+        if (existingUser != null) {
+            throw new IllegalStateException("User already exists");
+        }
+
         return userRepository.update(user);
     }
 
     public boolean deleteUserById(int id) {
+        UserEntity possibleUser = userRepository.findById(id);
+        if (possibleUser == null) {
+            throw new EntityNotFoundException("User not found");
+        }
+
         return userRepository.delete(id);
     }
 }
